// dev/spa-app/src/types/JamboApi.ts
export interface JamboRequest {
  readonly api: string;
}

export type JamboResponseData = {};

export type INVALID_PARAMETER = -2;
export type ERROR_IN_DB = -1;
export type UNKNOWN_ERROR = 1;
export type WRONG_DATA_FORMAT = 2;
export type INVALID_TOKEN = 3;
export type EMAIL_REGISTED = 12;
export type INVALID_USER_NAME = 14;
export type FAILED_ADD_REGISTER_BONUS = 17;
export type INVALID_PASSWORD = 21;
export type ALREADY_ADDED_BLOCK = 60;
export type NOT_ENOUGHT_POINT = 70;
export type EXIST_DATA = 75;
export type USER_NOT_EXIST = 80;
export type PARTNER_IS_NOT_WAITING_VIDEO_CALL = 82;
export type PARTNER_IS_NOT_WAITING_VOICE_CALL = 84;
export type PARTNER_IS_NOW_CALLING = 85;
export type PARTNER_IS_ALREADY_LIVE_NOW = 89;
export type ALREADY_ADDED_FAVORITE = 3117;
export type EMAIL_ALREADY_SENT = 8001;
export type LOVENSE_IS_OFFLINE = 5007;
export type INSUFFICIENT_POINTS = 5008;
export type LOVENSE_CONNECTION_ERROR = 5009;
export type LOVENSE_ROULETTE_LIMIT = 5010;
export type ALREADY_REGISTERED_WITH_GOOGLE = 9205;
export type VIDEO_FILE_NOT_FOUND = 71;
export type VIDEO_PERMISSION_DENIED = 72;
export type VIDEO_SIZE_EXCEEDED = 73;

type SUCCESS_CODE = 0;
type ERROR_CODE =
  | ERROR_IN_DB
  | INVALID_TOKEN
  | INVALID_PARAMETER
  | UNKNOWN_ERROR
  | WRONG_DATA_FORMAT
  | EMAIL_REGISTED
  | INVALID_USER_NAME
  | FAILED_ADD_REGISTER_BONUS
  | INVALID_PASSWORD
  | ALREADY_ADDED_BLOCK
  | NOT_ENOUGHT_POINT
  | USER_NOT_EXIST
  | ALREADY_ADDED_FAVORITE
  | EMAIL_ALREADY_SENT
  | LOVENSE_IS_OFFLINE
  | INSUFFICIENT_POINTS
  | LOVENSE_CONNECTION_ERROR
  | LOVENSE_ROULETTE_LIMIT
  | PARTNER_IS_NOT_WAITING_VOICE_CALL
  | PARTNER_IS_NOW_CALLING
  | PARTNER_IS_ALREADY_LIVE_NOW
  | EXIST_DATA
  | ALREADY_REGISTERED_WITH_GOOGLE
  | VIDEO_FILE_NOT_FOUND
  | VIDEO_PERMISSION_DENIED
  | VIDEO_SIZE_EXCEEDED;

export type JamboResponse<T extends JamboResponseData | JamboResponseData[]> =
  | {
      readonly code: SUCCESS_CODE;
      readonly data: T;
    }
  | {
      readonly code: ERROR_CODE;
      readonly data?: undefined;
    };